TST: Move plotting related tests to tests/plotting
