+++ setup.py
@@ -572,6 +572,7 @@ setup(name=DISTNAME,
                 'pandas.tests.formats',
                 'pandas.tests.types',
                 'pandas.tests.test_msgpack',
+                'pandas.tests.plotting',
                 'pandas.tools',
                 'pandas.tools.tests',
                 'pandas.tseries',
